перенес получение оффера в хук
