+++ src/pages/offer-page/offer-page.tsx
@@ -1,4 +1,5 @@
 import { useParams } from 'react-router-dom';
+import { useOffer } from '../../hooks/use-offer/use-offer';
 import { useCities } from '../../hooks/use-cities/use-cities';
 import Header from '../../components/header/header';
 import OfferGallery from '../../components/offer-gallery/offer-gallery';
@@ -6,14 +7,10 @@ import OfferContainer from '../../components/offer-container/offer-container';
 import PlaceCard from '../../components/place-card/place-card';
 import MapContainer from '../../components/map-container/map-container';
 import { OfferComment } from '../../types/offer-type';
-// import { IncomingOffer } from '../../types/offer-type';
 import { Offer } from '../../types/offer-type';
 import { RATING, Sign } from '../../const';
 import { Helmet } from 'react-helmet-async';
-import { fetchIncommingOffer } from '../../store/offers';
-import { useAppDispatch, useAppSelector } from '../../hooks/store/store';
-import { useEffect } from 'react';
-// import { useOffer } from '../../hooks/use-offer/use-offer';
+
 
 type OfferPageProps = {
   // incomingOffer: IncomingOffer;
@@ -27,15 +24,8 @@ type OfferPageProps = {
 function OfferPage ({sign, offerComments, rating, isOfferCard = true, selectedPoint}: OfferPageProps):JSX.Element {
 
   const { id } = useParams();
-  const dispatch = useAppDispatch();
 
-  useEffect(() => {
-    if (id) {
-      dispatch(fetchIncommingOffer({id}));
-    }
-  },[dispatch, id]);
-
-  const incomingOffer = useAppSelector((state) => state.offers.incommingOffer);
+  const { incomingOffer } = useOffer(id);
 
   const {activeOffers} = useCities();
   return (
